Add CubeServices

+++ frontend/src/services/room.service.js
@@ -15,27 +15,38 @@ export function removeRoom(roomId) {
 }
 
 export function updateRoom(roomData) {
-    return axios.patch(`${API_URL}/${roomData}`)
+    return axios.patch(`${API_URL}/${roomData}`);
 }
 
 export function fetchRoomById(roomId) {
-    return axios.get(`${API_URL}/${roomId}`)
+    return axios.get(`${API_URL}/${roomId}`);
 }
 
 export function joinRoom(roomId, virtualUser) {
-    return axios.post(`${API_URL}/${roomId}/users`, virtualUser)
+    return axios.post(`${API_URL}/${roomId}/users`, virtualUser);
 }
 
 export function leaveRoom(roomId, virtualUser, userId) {
-    if (userId || userId === 0)
-        return axios.delete(`${API_URL}/${roomId}/users/${roomId}`, virtualUser)
-    return axios.delete(`${API_URL}/${roomId}/users`, virtualUser)
+    if (userId || userId === 0) return axios.delete(`${API_URL}/${roomId}/users/${roomId}`, virtualUser);
+    return axios.delete(`${API_URL}/${roomId}/users`, virtualUser);
 }
 
 export function joinTeam(roomId, virtualTeam) {
-    return axios.post(`${API_URL}/${roomId}/teams`, virtualTeam)
+    return axios.post(`${API_URL}/${roomId}/teams`, virtualTeam);
 }
 
 export function leaveTeam(roomId, virtualTeam) {
-    return axios.delete(`${API_URL}/${roomId}/teams`, virtualTeam)
+    return axios.delete(`${API_URL}/${roomId}/teams`, virtualTeam);
+}
+
+export function connectPi(piName, roomId) {
+    return axios.post(`${API_URL}/${roomId}/connect_pi`, piName)
+}
+
+export function disconnectPi(piName, roomId) {
+    return axios.post(`${API_URL}/${roomId}/disconnect_pi`, piName)
+}
+
+export function searchCube(roomId) {
+    return axios.post(`${API_URL}/${roomId}/searchCube`)
 }
